Use LiveView form for internal performance creation
